[feat] TodoDetailScene의 mode에 따라 (.view, .new, .edit) todoDetailView의 userInteractionEnabled를 처리
