--- Planus/Presentation/TodoDetailScene2/ViewModel/TodoDetailViewModel2.swift
@@ -1,9 +1,0 @@
-//
-//  TodoDetailViewModel2.swift
-//  Planus
-//
-//  Created by Sangmin Lee on 2023/08/23.
-//
-
-import Foundation
-
